Rename involvementService.js to commentService.js
